requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
